Prod 환경 CI/CD 테스트

+++ src/main/java/com/kakaotech/ott/ott/post/domain/model/Post.java
@@ -30,7 +30,7 @@ public class Post {
     private Long viewCount = 0L;
     private int scrapCount;
 
-    private int weight;
+    private double weight;
 
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
@@ -45,7 +45,7 @@ public class Post {
                 .likeCount(0)
                 .viewCount(0L)
                 .scrapCount(0)
-                .weight(0)
+                .weight(0.0)
                 .build();
 
     }
